modify Gradient text components
